Changing from Tuple Struct to normal struct
Kept making too many mistakes forgetting which value was min and which
was max, so decided to make it clearer.

This does break compatibility with `Wrapping<T>`, where you could still
access the inner value with `.0` but it helps prevent many bugs which is
preferable.

People can do a find/replace for `s/\.0/.value/g` if they relly need to